refactor: :recycle: Clean up responsibilitites and internal logic of WaveFrontMaps.
